コメントの透明度変更 #22

+++ SRNicoNico/Models/NicoNicoWrapper/NicoNicoMylist.cs
@@ -14,7 +14,6 @@ using SRNicoNico.ViewModels;
 namespace SRNicoNico.Models.NicoNicoWrapper {
     public class NicoNicoMylist : NotificationObject {
 
-
         //とりあえずマイリスト取得API
         private const string DefListAPI = "http://www.nicovideo.jp/api/deflist/list";
 
@@ -70,9 +69,9 @@ namespace SRNicoNico.Models.NicoNicoWrapper {
                 //動画
                 if(data.Type == 0) {
 
-                    data.UpdateTime = UnixTime.FromUnixTime((long)item.update_time).ToString();
-                    data.FirstRetrieve = UnixTime.FromUnixTime((long)item.first_retrieve).ToString();
-                    data.Length = NicoNicoUtil.ConvertTime(long.Parse(item.length_seconds));
+                    data.UpdateTime = UnixTime.FromUnixTime((int)item.update_time).ToString();
+                    data.FirstRetrieve = UnixTime.FromUnixTime((int)item.first_retrieve).ToString();
+                    data.Length = NicoNicoUtil.ConvertTime(int.Parse(item.length_seconds));
                     data.Id = item.video_id;
                     data.ViewCounter = int.Parse(item.view_counter);
                     data.CommentCounter = int.Parse(item.num_res);
@@ -81,8 +80,8 @@ namespace SRNicoNico.Models.NicoNicoWrapper {
 
                 } else if(data.Type == 5) { //静画
 
-                    data.UpdateTime = UnixTime.FromUnixTime((long)item.update_time).ToString();
-                    data.FirstRetrieve = UnixTime.FromUnixTime((long)item.create_time).ToString();
+                    data.UpdateTime = UnixTime.FromUnixTime((int)item.update_time).ToString();
+                    data.FirstRetrieve = UnixTime.FromUnixTime((int)item.create_time).ToString();
                     data.Id = item.id.ToString();
                     data.ViewCounter = (int)item.view_count;
                     data.CommentCounter = (int)item.comment_count;
@@ -91,8 +90,8 @@ namespace SRNicoNico.Models.NicoNicoWrapper {
 
                 } else if(data.Type == 6) { //書籍
 
-                    data.UpdateTime = UnixTime.FromUnixTime((long)entry.update_time).ToString();
-                    data.FirstRetrieve = UnixTime.FromUnixTime((long)item.released_at).ToString();
+                    data.UpdateTime = UnixTime.FromUnixTime((int)entry.update_time).ToString();
+                    data.FirstRetrieve = UnixTime.FromUnixTime((int)item.released_at).ToString();
                     data.Id = "bk" + item.id;
                     data.ViewCounter = (int)item.view_count;
                     data.CommentCounter = (int)item.comment_count;
@@ -101,8 +100,8 @@ namespace SRNicoNico.Models.NicoNicoWrapper {
 
                 } else if(data.Type == 13) { //ブロマガ
 
-                    data.UpdateTime = UnixTime.FromUnixTime((long)item.commented_time).ToString();
-                    data.FirstRetrieve = UnixTime.FromUnixTime((long)item.create_time).ToString();
+                    data.UpdateTime = UnixTime.FromUnixTime((int)item.commented_time).ToString();
+                    data.FirstRetrieve = UnixTime.FromUnixTime((int)item.create_time).ToString();
                     data.Id = "ar" + item.id;
                     data.CommentCounter = (int)item.comment_count;
                     data.MylistCounter = int.Parse(item.mylist_count);
